{rolling} Rename demo apps to avoid conflicts
The ROS demo and examples packages are usually saved in custom directories
in the format: /opt/ros/<DISTRO>/lib/<PACKAGENAME>/<APP>

Since we install everything into the main filesystem hierarchy this means
that applications with the same name collide. (such as listener)

We prefix the executables with the package name to help differentiate them.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>