scripts/ros-cherrypick.sh: Tool for porting commits between ROS distro layers
Automates the porting of patches between ROS distro layers within the same
Yocto release / branch. It iterates over the commits in a specified range and automatically
renames the Bitbake recipe files based on the versions found in the new ROS distro layer.

It also automatically filters out files from ROS packages that no longer exist.  If a patch
file no longer contains any changes, the commit is discarded.

The result is a series of patch files created in the local directory based on each commit.
One may make any necessary changes to the commit messages or additional fixes for things
that cannot be automated.  When ready, the patches may be applied with git-am.

  eg. scripts/ros-cherrypick.sh meta-ros2-humble meta-ros2-iron 791990e9 28e09796

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>