{rolling} lely-core: Add new recipe for lely-core-libraries
The lely-core-libraries package from ROS needs lely-core and tries
to pull it in with ExternalProject_Add.

However, it uses CMake and lely-core uses autotools.  This makes it
difficult to neatly cross-compile.  Instead, it is easier to break
out lely-core as a separate recipe and let lely-core-libraries
have a dependency on it.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>