{rolling} lely-core-libraries: Fix cross-compilation problems
The lely-core-libraries package depends on lely-core, however the
former uses CMake and the latter uses GNU Autotools.

In order to do a clean cross-build, the lely-core build has been
broken out into a separate recipe. A dependency on lely-core-native
is needed for the build tools.  A dependency on lely-core is
needed to provide liblely libraries at runtime.

Fixes are also applied to allow the build to be done with GNU Make
and fix the Python script shebangs.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>