{rolling} packagegroup-ros-world-rolling: Remove broken packages
These packages are failing to build and are likely non-trivial to fix.
We are removing them from the world build to allow us to move forward
with continuous integration builds.

smacc2 has been failing rolling builds on the ROS 2 Buildfarm since June
2023.

The novatel-gps-driver is specific to a particular GPS device and may be
fixed if someone using the driver comes forward.

The canopen-tests recipe creates a dcf-tools wheel that does not get
installed into the destination directory.

The nao-lola has been deprecated in favour of nao-lola-client.
Attempting to install both results in a conflict as they provide some
of the same files.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>