{common}{humble}{iron}{rolling} cmake: Remove recipe
Copies of the cmake recipe were added for older Yocto Project
releases that did not meet the minimum version requirement.
This is no longer required by any supported releases and the
one provided by OpenEmbedded is newer.

Also remove the cmake bbappend in meta-ros-common.  It was
added originally to prevent the toolchain file for the SDK
from being placed in nativesdk-cmake-dev.  However, this
no longer happens and the change is no longer required.

Signed-off-by: Matthias Schoepfer <matthias.schoepfer@googlemail.com>
Signed-off-by: Rob Woolley <rob.woolley@windriver.com>