{humble} Revert Rename demo apps to avoid conflicts
The ROS demo and examples packages are usually saved in custom
directories in the format:
    /opt/ros/<DISTRO>/lib/<PACKAGENAME>/<APP>

However, they were being installed in /usr/bin due to changes to
the setuptools3 class.  Adding a ${PN} prefix to the files was a
temporary workaround to prevent packages from providing files with
the same name.

The install location has been resolved in ros_ament_python.bbclass
to use setuptools3_legacy instead. The workaround is no longer
necessary.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>