{rolling} fixing orocos-kdl-vendor
The recipe was failing with
```
Fetcher failure: Unable to find revision f1bbaed1b5190dbfefd08ca445e53d20d1a0066e in branch release/rolling/orocos_kdl_vendor
```
This comes from the patch that will be dependent on the hashes of the package,
therefore this approach is prone to fail on every refresh of this package.

The alternative approach is to introduce a yocto recipe for orocos-kdl, which
basically makes any appends to orocos-kdl-vendor obsolete.

orocos-kdl can be used in other distros (humble, noetic) as well, so I placed
it into ros-common layer.

Signed-off-by: Matthias Schoepfer <matthias.schoepfer@googlemail.com>