{rolling} fix build of ecl-license
This fixes two issues. The first one is that the package.xml is
listening the wrong SPDX (BSD instead of BSD-3-Clause). The second one
is that this package installs license files (into the wrong directory btw.)
So to fix this properly, we only care for those files *iff* we build
license packages. And if we do, we move them to the right directory.

Signed-off-by: Matthias Schoepfer <matthias.schoepfer@googlemail.com>