{common} jasper: Set the math library name
We can't use CMake's find_library() macro to find libm in OpenEmbedded
when building native recipes.  CMake is intentionally configured
to not look in the host library paths.

As a result, we must ensure that CMake is provided with a suitable
default value for the name of the math library.

Upstream-Status: Pending

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>