meta-ros-common: Add ros_optdir to ros_opt_prefix.bbclass
Many packages put extra files under /opt/ros/<DISTRO>/opt/<packagename>.

This includes the Gazebo vendor packages (eg. gz-math-vendor and
sdformat-vendor) which are required to build packages that depend on
them (eg. sdformat-urdf).

Add this path to SYSROOT_DIRS to ensure that they are included in
the recipe sysroot and can be found in do_configure.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>