meta-ros-common: ros_opt_prefix.bbclass: Add ROS opt prefix .so symlinks
The FILES_SOLIBSDEV variable is used to specify which shared library
symlinks should be packages in the -dev packages.

We must add the ROS opt prefix library path to this variable instead
of directly in the FILES variable defaults. This allows
ros_faulty_solibs.bbclass to still override the inclusion of the
files when shared libraries are missing their version suffix.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>