{rolling} mola-traj-tools: Modify shell script shebang
This resolves an error raised during the do_package_qa()
task where it detects a missing runtime dependency on /bin/env.

Adding a hardcoded, runtime dependency on coreutils or busybox is
not ideal as it removes the flexibility to choose.

The bitbake file-rdeps code has an exception for /usr/bin/env.
Thus, by changing the shebang we eliminate the error.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>