{rolling} mrpt2: Resolve build failures
Add a missing = prefix to isystem to use the sysroot directory as the
base.  Backport an upstream fix for a missing deconstructor symbol.
Disable using CMAKE_FIND_ROOT_PATH when searching for a font in the same
directory.  Set the CMAKE_TOOLCHAIN_FILE for a subproject to ensure that
it uses the correct compiler configuration.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>