{rolling} qt-gui-cpp: Conditionally inherit classes
Bitbake will fail to execute if the meta-qt5 layer is not included and
the inherited class files cannot be found.

This change makes inheriting the classes conditional so that bitbake may
parse the recipe.  There is a SKIP_RECIPE in place that will avoid
building the recipe if it is attempted without the meta-qt5 layer.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>