meta-ros1: ros_catkin.bbclass: Remove host paths
Catkin provides a template file (pkgConfig.cmake.in) with a section of
code that is unreachable on the target. This section references
CMAKE_CURRENT_SOURCE_DIR which causes host paths to leak into generated
CMake files. This causes do_package_qa to fail on ROS packages that use
catkin.

We can safely remove the host paths when building for the target without
making any changes to the end result

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>