{common} cpptoml: Fix failure to find libc++abi when compiling with LLVM/Clang
Downstream packages like iceoryx-posh were failing to build because cpptoml
was exporting absolute paths from the host in the CMake target
configuration files. This was caused by a condition that used MeTA-cmake
to find libcxx when Clang is used.  Removing this conditional block allows
cpptoml to build for both GCC and Clang.

Upstream-Status: Inactive-Upstream

Signed-off-by: Alexander Entinger <consulting@lxrobotics.com>