{noetic}{humble}{iron}{jazzy}{rolling} ros-core: Restore dependencies
The ros-core variant is defined to have runtime dependencies that are
required for building ROS.  In the past when images were assumed to
no support on-target development it made sense to remove them.

However, this makes the ros-core package dependencies in OpenEmbedded
different than ROS 2 on other platforms.  The ros-core metapackage
should match ros_core on Ubuntu to follow the principle of least
surprise.

If someone wants to minimize the footprint they should
add the packages they want individually rather than try to remove them
from ros-core.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>