{common} libglvnd: add new recipe libglvnd v1.6.0
This is a backport of the recipe from Mickledore.  It is needed
to properly detect OpenGL using FindOpenGL.cmake.

This recipe is meant to be an optional dependency of mesa.
This recipe will allow for glvnd support in mesa. Thus,
giving access to the GLX window-system API library and
related libraries, mesa will build libGLX_mesa.so.*.*
and libEGL_mesa.so.*.* or whatever specified in
-Dglvnd-vendor-name. There may be applications that require
direct access to the GLX window-system API library in order
to build/execute. This will allow for multiple layers to
have generic support for libraries in libglvnd
(GLX window-system API library, etc...) without needing to
implement their own version of the libglvnd recipe.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>
Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Wang Mingyu <wangmy@fujitsu.com>
Signed-off-by: Rob Woolley <rob.woolley@windriver.com>