{rolling} ament-package: Enable compatibility with busybox
The existing implementation used options to head and tail that
weren't present in busybox by default. This caused problems
when trying to source /opt/ros/rolling/setup.sh

It is simpler to just remove a leading or trailing colon. This has the
added benefit of only using shell built-in functions.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>