generate_recipes.yml: Add support for pre-released ROS distros
When a new ROS distribution is being created there is an extended period
of time between when the named directory is created in rosdistro and
when it is tagged.

During this time we must pass in the value "pre-release" into
the generate recipe shell scripts.  We also should use the
latest commit in the repository.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>