{common} gazebo11: Disable build-time checks for binaries
The CMake target for hdf5 attempts to check that all libraries and
executables are present.  However, Yocto filters out target executables
when installing files in the recipe-sysroot so the check fails.

We remove those lines checking for files in the bin directory as
they are not needed for the build.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>