kas: Append mirror_updates to end of INHERIT
The sstate cache was failing to mirror to the S3 buckets.

For mirror_updates.bbclass to properly initialize it must be
added to INHERIT after sstate.bbclass. Even though += was used
in local.conf it was parsed before the INHERIT changes made by
defaultsetup.conf

This caused mirror_updates.bbclass to initialize first and
miss the sstate_task_postfunc varflags that were added later
to the SSTATETASKS by sstate.bbclass.

Using the append override defers modifications to INHERIT
until the end of parsing.

Moving the configuration to auto.conf or layer.conf may
be an alternative solution.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>