ros-generate-recipes.sh: Specify Yocto release
The ros-generate-recipes.sh script operates on an individual ROS distro
layer. Use the layer configuration file from the layer to identify the
target Yocto Project release.

Use this information to explicitly set --yocto-release when calling
superflore-gen-oe-recipes.  This allows the superflore tool to adjust the
generated recipe based on differences in newer releases.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>