{common} libffi: Build with PIC support
Under certain conditions the libffi.a static library is built
with non-relocatable symbols in it. This causes openmp builds
to fail.

This is a known issue with Linux distributions that package libffi.
Debian provides a libffi_pic.a specifically for this purpose.

For the time-being we will work around the problem by making
sure that libffi is built with --with-pic when using meta-ros.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>