ros-generate-recipes: Specify Yocto release
Recent changes in Whinlatter requires that the Yocto release
be supplied to superflore to generate recipe correctly.

The ros-generate-recipes.sh script operates on individual
layers which are compatible with a given Yocto Project release.

These changes use the layer configuration file to detect
the release and supply it to the superflore command.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>