Update LAYERSERIES_COMPAT to kirkstone
As of the Yocto 4.2 release, the LAYERSERIES_COMPAT variable in
layer.conf no longer expands variables. This was done to make it clearer
which Yocto releases are supported by a given layer.

See the Yocto Release 4.2 (mickledore) migration guide for details:
  https://docs.yoctoproject.org/next/migration-guides/migration-4.2.html

Kirkstone is the last release prior to this change. Making this change
keeps it consistent with the later releases.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>